fix(agent,schema): Windows exit-code fidelity, alignment, idempotent stop
Copilot re-review of the ETW sensor:
- sensor: emit the Windows exit code as an unsigned DWORD widened to int, not
  through int32, so NTSTATUS/HRESULT crash codes (0xC0000005 etc.) are not
  flipped into misleading negatives.
- etw: decode UTF-16 string properties via encoding/binary instead of
  reinterpreting the []byte as []uint16 through unsafe (unaligned view can fault
  or misread on ARM64), matching the uint32/int64 decode paths.
- etw: make Session.Stop idempotent by treating ERROR_WMI_INSTANCE_NOT_FOUND as
  already-stopped, so teardown does not warn on the normal double-stop path (now
  that Disconnect logs Stop errors).
- schema: bound create_time_ns with "minimum": 0 on exec_payload and
  exit_payload, matching pidversion (it is a Unix-ns timestamp, never negative).

+++ agent/wintel/sensor_windows.go
@@ -164,9 +164,11 @@ func (s *Sensor) exitFromRecord(r etw.Record) ([]byte, error) {
 		return nil, fmt.Errorf("etw exit: missing required fields (pid=%d okPID=%v createTime=%d okCT=%v)", pid, okPID, createFT, okCT)
 	}
 	exitCode, _ := r.Uint32("ExitCode") // best-effort: 0 is a legitimate (success) exit code, so a missing/zero value is not fatal
+	// A Windows exit code is an unsigned DWORD; NTSTATUS/HRESULT-style crash codes (e.g. 0xC0000005) have the high bit set. Widen the
+	// uint32 straight to int (non-negative on 64-bit) rather than through int32, which would flip those into misleading negatives.
 	return exitEnvelope(uuid.NewString(), s.hostID, time.Now().UnixNano(), exitPayload{
 		PID:          int(pid),
-		ExitCode:     int(int32(exitCode)),
+		ExitCode:     int(exitCode),
 		CreateTimeNs: filetimeToUnixNano(createFT),
 	})
 }
